support errors.Is(err, MyClass.Instance())
The MyClass.Has(err) idiom has never been anyone's favorite thing. It
was a necessary compromise given what was available to us at the time.
But with the help of errors.Is(), maybe something like this is better?

    MyClass := errs.Class("foo")
    err := MyClass.Wrap(io.EOF)
    if errors.Is(err, MyClass.Instance()) {
        fmt.Println("err is an instance of MyClass")
    }

It's still a little clunky, since we have to have the .Instance() method
there at all. I think it's necessary to avoid having Class implement the
error interface itself, and I think that's always been an anti-goal.

Suggestions welcome for a better name.