midstack inline wrap and precice allocate pcs
this makes it so that doing

    if err != nil {
        return errs.Wrap(err)
    }

should avoid double checking the err for nilness.

additionally, this stack allocates the buffer for
the pcs to put the callers into before heap allocating
the exact size necessary. this reduces the memory usage
an error with a small traceback by like 5x or more.